Create Local Interactions _ 5XS2
This is the commands for local interactions on 5XS2, CDK8-CycC run with mutatex, as well as the commands for visualisation, CDK8 is chain A and CycC is chain B. A seperate file called posnew.txt is provided with all the mutations for this step. The same file is used for Mutatex for the Ensemble on local interactions in the Models.